requirements.txt
discord.py==2.5.2
python-dotenv==0.21.0
discord-webhook==1.0.0
lz==2.0.0